Reject table scan for INDEXED format or KV table
