+++ crates/fluss/src/client/table/scanner.rs
@@ -34,8 +34,9 @@ use crate::client::table::log_fetch_buffer::{
     LogFetchBuffer, RemotePendingFetch,
 };
 use crate::client::table::remote_log::{RemoteLogDownloader, RemoteLogFetchInfo};
+use crate::error::Error::UnsupportedOperation;
 use crate::error::{ApiError, Error, FlussError, Result};
-use crate::metadata::{PhysicalTablePath, TableBucket, TableInfo, TablePath};
+use crate::metadata::{LogFormat, PhysicalTablePath, TableBucket, TableInfo, TablePath};
 use crate::proto::{ErrorResponse, FetchLogRequest, PbFetchLogReqForBucket, PbFetchLogReqForTable};
 use crate::record::{
     LogRecordsBatches, ReadContext, ScanBatch, ScanRecord, ScanRecords, to_arrow_schema,
@@ -107,7 +108,7 @@ impl<'a> TableScan<'a> {
     ///     let table = conn.get_table(&table_path).await?;
     ///
     ///     // Project columns by indices
-    ///     let scanner = table.new_scan()?.project(&[0, 2, 3])?.create_log_scanner()?;
+    ///     let scanner = table.new_scan().project(&[0, 2, 3])?.create_log_scanner()?;
     ///     let scan_records = scanner.poll(Duration::from_secs(10)).await?;
     ///     for record in scan_records {
     ///         let row = record.row();
@@ -182,7 +183,7 @@ impl<'a> TableScan<'a> {
     ///     let table = conn.get_table(&table_path).await?;
     ///
     ///     // Project columns by column names
-    ///     let scanner = table.new_scan()?.project_by_name(&["col1", "col3"])?.create_log_scanner()?;
+    ///     let scanner = table.new_scan().project_by_name(&["col1", "col3"])?.create_log_scanner()?;
     ///     let scan_records = scanner.poll(Duration::from_secs(10)).await?;
     ///     for record in scan_records {
     ///         let row = record.row();
@@ -221,6 +222,7 @@ impl<'a> TableScan<'a> {
     }
 
     pub fn create_log_scanner(self) -> Result<LogScanner> {
+        validate_scan_support(&self.table_info.table_path, &self.table_info)?;
         let inner = LogScannerInner::new(
             &self.table_info,
             self.metadata.clone(),
@@ -234,6 +236,7 @@ impl<'a> TableScan<'a> {
     }
 
     pub fn create_record_batch_log_scanner(self) -> Result<RecordBatchLogScanner> {
+        validate_scan_support(&self.table_info.table_path, &self.table_info)?;
         let inner = LogScannerInner::new(
             &self.table_info,
             self.metadata.clone(),
@@ -1556,6 +1559,25 @@ impl BucketScanStatus {
     }
 }
 
+fn validate_scan_support(table_path: &TablePath, table_info: &TableInfo) -> Result<()> {
+    if table_info.schema.primary_key().is_some() {
+        return Err(UnsupportedOperation {
+            message: format!("Table {table_path} is not a Log Table and doesn't support scan."),
+        });
+    }
+
+    let log_format = table_info.table_config.get_log_format()?;
+    if LogFormat::ARROW != log_format {
+        return Err(UnsupportedOperation {
+            message: format!(
+                "Scan is only supported for ARROW format and table {table_path} uses {log_format} format"
+            ),
+        });
+    }
+
+    Ok(())
+}
+
 #[cfg(test)]
 mod tests {
     use super::*;
@@ -1564,7 +1586,7 @@ mod tests {
     use crate::compression::{
         ArrowCompressionInfo, ArrowCompressionType, DEFAULT_NON_ZSTD_COMPRESSION_LEVEL,
     };
-    use crate::metadata::{PhysicalTablePath, TableInfo, TablePath};
+    use crate::metadata::{DataTypes, PhysicalTablePath, Schema, TableInfo, TablePath};
     use crate::record::MemoryLogRecordsArrowBuilder;
     use crate::row::{Datum, GenericRow};
     use crate::rpc::FlussError;
@@ -1780,5 +1802,76 @@ mod tests {
 
         assert!(metadata.get_cluster().leader_for(&bucket).is_none());
         Ok(())
+    }
+
+    fn create_test_table_info(
+        has_primary_key: bool,
+        log_format: Option<&str>,
+    ) -> (TableInfo, TablePath) {
+        let mut schema_builder = Schema::builder()
+            .column("id", DataTypes::int())
+            .column("name", DataTypes::string());
+
+        if has_primary_key {
+            schema_builder = schema_builder.primary_key(vec!["id"]);
+        }
+
+        let schema = schema_builder.build().unwrap();
+        let table_path = TablePath::new("test_db", "test_table");
+
+        let mut properties = HashMap::new();
+        if let Some(format) = log_format {
+            properties.insert("table.log.format".to_string(), format.to_string());
+        }
+
+        let table_info = TableInfo::new(
+            table_path.clone(),
+            1,
+            1,
+            schema,
+            vec![],
+            Arc::from(vec![]),
+            1,
+            properties,
+            HashMap::new(),
+            None,
+            0,
+            0,
+        );
+
+        (table_info, table_path)
+    }
+
+    #[test]
+    fn test_validate_scan_support() {
+        // Primary key table
+        let (table_info, table_path) = create_test_table_info(true, Some("ARROW"));
+        let result = validate_scan_support(&table_path, &table_info);
+
+        assert!(result.is_err());
+        let err = result.unwrap_err();
+        assert!(matches!(err, UnsupportedOperation { .. }));
+        assert!(err.to_string().contains(
+            format!("Table {table_path} is not a Log Table and doesn't support scan.").as_str()
+        ));
+
+        // Indexed format
+        let (table_info, table_path) = create_test_table_info(false, Some("INDEXED"));
+        let result = validate_scan_support(&table_path, &table_info);
+
+        assert!(result.is_err());
+        let err = result.unwrap_err();
+        assert!(matches!(err, UnsupportedOperation { .. }));
+        assert!(err.to_string().contains(format!("Scan is only supported for ARROW format and table {table_path} uses INDEXED format").as_str()));
+
+        // Default format
+        let (table_info, table_path) = create_test_table_info(false, None);
+        let result = validate_scan_support(&table_path, &table_info);
+        assert!(result.is_ok());
+
+        // Arrow format
+        let (table_info, table_path) = create_test_table_info(false, Some("ARROW"));
+        let result = validate_scan_support(&table_path, &table_info);
+        assert!(result.is_ok());
     }
 }
